app/requirements.txt
fastapi==0.96.0
sqlalchemy==2.0.15
sqlalchemy_utils==0.41.1
alembic==1.11.1
orjson==3.9.0
pydantic==1.10.8
uvicorn==0.22.0
uvloop==0.17.0
gunicorn==20.1.0
asyncpg==0.27.0
psycopg2-binary==2.9.6
aiohttp[speedups]
python-multipart==0.0.6
python-magic==0.4.27
pydub==0.25.1
python-dotenv==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==4.5.5